# bc7enc: Optimize "find approximate selector" branch chains